+++ config/urls.py
@@ -9,22 +9,12 @@ from django.contrib import admin
 from django.views.generic import TemplateView
 from django.views import defaults as default_views
 
-# Third party imports
-from filebrowser.sites import site
-
-# The filebrowser app needs to be a sub path of the admin url, and since the admin url is
-# not static, we have to assemble this sub path here first using the admin url variable
-FILEBROWSER_URL = os.path.join(settings.ADMIN_URL, 'filebrowser')
-
 
 urlpatterns = [
     path("", TemplateView.as_view(template_name="pages/home.html"), name="home"),
     path(
         "about/", TemplateView.as_view(template_name="pages/about.html"), name="about"
     ),
-    # Config for the filebrowser app, which is needed for the editor. This needs to be added before the base
-    # admin url is defined.
-    path(FILEBROWSER_URL, site.urls),
     # Django Admin, use {% url 'admin:index' %}
     path(settings.ADMIN_URL, admin.site.urls),
     # User management
@@ -32,9 +22,9 @@ urlpatterns = [
     path("accounts/", include("allauth.urls")),
     # Your stuff: custom urls includes go here
     path("blog/", include("pitronically.blog.urls", namespace="blog")),
-    # TinyMCE Editor for blog posts
-    #path("tinymce/", include("tinymce.urls")),
-    path("ckeditor/", include("ckeditor_uploader.urls"))
+    # CKEditor Editor for blog posts
+    path("ckeditor/", include("ckeditor_uploader.urls")),
+    path("filer/", include("filer.urls")),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 
 if settings.DEBUG: